redesign auth AdminController

+++ project_v2/project_v2/project_v2/Controllers/AdminController.cs
@@ -4,25 +4,28 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Caching.Distributed;
 using Newtonsoft.Json;
 using project_v2.Models;
 using project_v2.Services.Interface;
 
 namespace project_v2.Controllers
 {
-    [Authorize(Roles = "Admin")]
     public class AdminController : Controller
     {
         private IAccountService accountsvc;
         private IRankService ranksvc;
         IServiceConfigurations serviceConfig;
         private IStatusService statussvc;
+        public IDistributedCache cache;
 
         public AdminController(IRankService _ranksvc,
             IStatusService _statussvc,
             IAccountService _accountsvc,
-            IServiceConfigurations serviceConfig)
+            IServiceConfigurations serviceConfig,
+            IDistributedCache _cache)
         {
+            cache = _cache;
             this.accountsvc = _accountsvc;
             this.ranksvc = _ranksvc;
             this.statussvc = _statussvc;
@@ -31,6 +34,11 @@ namespace project_v2.Controllers
 
         public IActionResult AccountManage()
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if(!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             var model = accountsvc.GetAllAccount();
             return View(model);
         }
@@ -38,6 +46,11 @@ namespace project_v2.Controllers
         [HttpGet]
         public IActionResult EditAccount(string id)
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             var user = accountsvc.GetAllAccount().FirstOrDefault(it => it._id == id);
             if (user == null) return RedirectToAction(nameof(AccountManage));
             var json = JsonConvert.SerializeObject(user);
@@ -63,6 +76,11 @@ namespace project_v2.Controllers
         [HttpPost]
         public IActionResult EditAccount(EditAccountModel model, string id)
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             if (ModelState.IsValid)
             {
                 var allAcc = accountsvc.GetAllAccount();
@@ -87,6 +105,11 @@ namespace project_v2.Controllers
 
         public IActionResult ProjectSystem()
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             ViewBag.GuestRankId = serviceConfig.GuestRankId;
             ViewBag.MasterRankId = serviceConfig.MasterRankId;
             ViewBag.NewStatusId = serviceConfig.StatusNewId;
@@ -100,12 +123,22 @@ namespace project_v2.Controllers
 
         public IActionResult CreateRank()
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             return View();
         }
 
         [HttpPost]
         public IActionResult CreateRank(ProjectRankModel model)
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             if (ModelState.IsValid)
             {
                 ranksvc.AddRank(model);
@@ -116,6 +149,11 @@ namespace project_v2.Controllers
 
         public IActionResult EditRank(string rankid)
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             var rankList = ranksvc.GetAllRank();
             var result = rankList.FirstOrDefault(it => it._id == rankid);
             return View(result);
@@ -124,6 +162,11 @@ namespace project_v2.Controllers
         [HttpPost]
         public IActionResult EditRank(ProjectRankModel model)
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             if (ModelState.IsValid)
             {
                 ranksvc.EditRank(model);
@@ -134,18 +177,33 @@ namespace project_v2.Controllers
 
         public IActionResult DeleteRank(string rankid)
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             ranksvc.DeleteRank(rankid);
             return RedirectToAction(nameof(ProjectSystem));
         }
 
         public IActionResult CreateStatus()
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             return View();
         }
 
         [HttpPost]
         public IActionResult CreateStatus(StatusModel model)
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             if (ModelState.IsValid)
             {
                 statussvc.AddStatus(model);
@@ -156,6 +214,11 @@ namespace project_v2.Controllers
 
         public IActionResult EditStatus(string statusid)
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             var statusList = statussvc.GetAllStatus();
             var result = statusList.FirstOrDefault(it => it._id == statusid);
             return View(result);
@@ -164,6 +227,11 @@ namespace project_v2.Controllers
         [HttpPost]
         public IActionResult EditStatus(StatusModel model)
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             if (ModelState.IsValid)
             {
                 statussvc.EditStatus(model);
@@ -174,6 +242,11 @@ namespace project_v2.Controllers
 
         public IActionResult DeleteStatus(string statusid)
         {
+            ViewBag.IsLogin = !string.IsNullOrEmpty(cache.GetString("user"));
+            if (ViewBag.IsLogin) ViewBag.User = JsonConvert.DeserializeObject<AccountModel>(cache.GetString("user"));
+            else return RedirectToAction("Login", "Account");
+            if (!ViewBag.User.IsAdmin) return RedirectToAction("Login", "Account");
+
             statussvc.DeleteStatus(statusid);
             return RedirectToAction(nameof(ProjectSystem));
         }
